[YS-176] refactor: funnel 컴포넌트 파일 분리
