+++ src/app/join/components/JoinInput/JoinInput.styles.ts
@@ -51,7 +51,6 @@ export const errorMessage = (theme: Theme) => css`
 export const tipWrapper = (theme: Theme) => css`
   ${theme.fonts.label.small.M12};
   display: flex;
-  // justify-content: flex-end;
   color: ${theme.colors.text02};
 `;
